### 2021-02-04 #3007 @whedon
[PRE REVIEW]: IGRINS RV: A Python Package for Precision Radial Velocities with Near-Infrared Spectra
**Submitting author:** @shihyuntang (<a href="http://orcid.org/0000-0003-4247-1401">Shih-Yun Tang</a>)
**Repository:** <a href="https://github.com/shihyuntang/igrins_rv" target ="_blank">https://github.com/shihyuntang/igrins_rv</a>
**Version:** v0.9.6-beta.1
