### 2021-02-04 #3007 @kyleniemeyer
Hi @shihyuntang, we use this pre-review issue to identify an editor and reviewers, and also fix any obvious issues.

Your paper is currently missing a Statement of Need section that clearly illustrates the research purpose of the software, which is one of [our required paper elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Can you please add this?

Also, our bot identified some potential missing DOIs; however, these may be for references in your `.bib` file but not actually included in the paper. Please confirm that all your references do have DOIs if available.
