### 2021-02-04 #3007 @kyleniemeyer
Hi @dfm, could you edit this submission for us?

I want to point out (for potential reviewers) that this submission is connected to an article submitted to the AAS Astronomical Journal. As part of JOSS's collaboration with AAS, they make a small donation to the running costs of JOSS.

AAS publishing, and that as part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS.

Some related links:
- Blog post describing the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
- Summary of JOSS financial model: https://joss.theoj.org/about#costs
