+++ paper/paper.md
@@ -39,6 +39,9 @@ The relative radial velocity of a star with respect to the Sun can be calculated
 
 In the past decade, RVs have primarily been measured from spectra in the optical wavelength regime. This is partly because of advancements in detector technology, but also because of the paucity of Earth's atmospheric absorption features (telluric lines) in the optical. Yet for a fainter, cooler, smaller stellar object like an M type star (stars with mass less than half of the Sun), which emits more energy in the Near-Infrared (NIR), observations in the NIR can save a considerable amount of exposure time. Also, the M type star is the most common type of star. This along with its size increases the detectability of Earth-like planets around them. Moreover, the stellar activity that can drive false positive exoplanet detections, e.g., star spots carried into view by stellar rotation is shown to be less severe in the NIR compared to optical.
 
+
+# Statement of need
+
 ``IGRINS RV`` is a pipeline built for extracting precision RVs from spectra taken with the Immersion GRating INfrared Spectrometer (IGRINS) spectrograph [@yuk10; @park14; @mace16; @mace18]. This pipeline is built on the forward-modeling methodology that was successfully applied to CSHELL and PHOENIX spectra in the past [@croc12]. Compared to RVs obtained via cross-correlation with stellar templates used by past studies, ``IGRINS RV`` gives three times higher RV precision, about 25--30 m/s, around narrowline stars in both H and K bands, shown by years of monitoring on two RV standard stars, GJ\ 281 and HD\ 26257. This improvement is because of the use of a more robust approach to wavelength calibration and better telluric modeling through the generation of high resolution synthetic templates from fits to A0 standard observations. ``IGRINS RV`` has also demonstrated its effectiveness in identifying orbiting companions by successfully recovering the planet-induced RV signals of HD\ 189733 and Tau\ Boo\ A. ``IGRINS RV`` lets users choose to obtain absolute RVs or relative RVs, depending on whether their priority is precise RV monitoring or more coarse RV characterization. ``IGRINS RV`` requires that the ``igrins plp v2.2.0`` [@leegul17] and ``Telfit`` [@gull14] packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.
 
 
