+++ paper/paper.md
@@ -50,6 +50,6 @@ In the past decade, RVs have primarily been measured from spectra in the optical
 
 # Acknowledgements
 
-Partial support for this work was provided by NASA Exoplanet Research Program grant 80-NSSC19K-0289 to L. Prato.  We are grateful for the generous donations of John and Ginger Giovale, the BF Foundation, and others which made the IGRINS-DCT program possible. Additional funding was provided by the Mt. Cuba Astronomical Foundation and the Orr Family Foundation. IGRINS was developed under a collaboration between the University of Texas at Austin and the Korea Astronomy and Space Science Institute (KASI) with the financial support of the US National Science Foundation under grant AST-1229522 and AST-1702267, of the University of Texas at Austin, and of the Korean GMT Project of KASI.
+Partial support for this work was provided by NASA Exoplanet Research Program grant 80-NSSC19K-0289 to L. Prato. CMJ would like to acknowledge partial support for this work provided through grants to Rice University provided by NASA (award 80-NSSC18K-0828) and the NSF (awards AST-2009197 and AST-1461918). We are grateful for the generous donations of John and Ginger Giovale, the BF Foundation, and others which made the IGRINS-DCT program possible. Additional funding was provided by the Mt. Cuba Astronomical Foundation and the Orr Family Foundation. IGRINS was developed under a collaboration between the University of Texas at Austin and the Korea Astronomy and Space Science Institute (KASI) with the financial support of the US National Science Foundation under grant AST-1229522 and AST-1702267, of the University of Texas at Austin, and of the Korean GMT Project of KASI.
 
 # References
