+++ Example/Prepdata_targ_GJ281.txt
@@ -1,4 +1,4 @@
-night beam tag mjd facility airmass bvc
+night beam tag jd facility airmass bvc
 20170106 A 0094 2457760.9221165 DCT 1.4 4.405562660442322
 20170106 B 0095 2457760.923446 DCT 1.41 4.4023485353486365
 20170128 A 0073 2457782.8031455 DCT 1.21 -6.333473651631627
