### 2021-02-22 #3007 @shihyuntang
Dear @dfm Thank you for your notification. We understand that. We also used this time to import our wiki page. Now there is a dedicate "Demo with Example Data" to work user through the entire process with the example data.
However, we fix some small bugs as well during the process.
Should I just make a new release like v0.9.6-beta.2 (submitted verse was v0.9.6-beta.1) or I need to make a new submission?
Thank you.
