### 2021-03-10 #3095 @whedon
[REVIEW]: IGRINS RV: A Python Package for Precision Radial Velocities with Near-Infrared Spectra
**Submitting author:** @shihyuntang (<a href="http://orcid.org/0000-0003-4247-1401">Shih-Yun Tang</a>)
**Repository:** <a href="https://github.com/shihyuntang/igrins_rv" target ="_blank">https://github.com/shihyuntang/igrins_rv</a>
**Version:** v1.0.0
