### 2021-04-02 shihyuntang/igrins_rv#2 @gully
Miscellaneous notes from JOSS review
Hi there :wave:
I'll be reviewing this package for JOSS, the key ideas of the review can be found at https://github.com/openjournals/joss-reviews/issues/3095.  The JOSS guidelines recommend keeping the details in Issues within this target repository, so I'll update this particular issue with realtime comments, and if the need arises will make standalone Issues as the review progresses.

## Preliminaries

### Repository size
The repository weighs in at 422 MB, with 209 MB in `Engine`, 50 MB in `Example`, and a 154 MB `.git`.  Within the `Engine` directory, the `syn_template` is 177 MB. The relatively large size stems in part from the inclusion of the example data and synthetic templates.  Have you considered the tradeoffs of storing the data and models separately and including a script to fetch them programmatically?

### Inclusion of raw TelFit code
The TelFit source code is included in `Telluric-Fitter-master`, with this note in the Wiki:
> Note that the original Telfit package is available on https://github.com/kgullikson88/Telluric-Fitter, but the files that come with IGRINS RV are modified, such that one line in ./src/TelluricFitter.py is different:

Is that one line a bug in TelFit?  If so, have you considered filing a patch?

### Citation of dependencies
The reference for the dependencies like [astropy](https://www.astropy.org/acknowledging.html), [numpy](https://numpy.org/citing-numpy/), and others may have to be included in the paper, see the JOSS guidelines:
> A list of key references, including to other software addressing related needs.
