### 2021-04-05 shihyuntang/igrins_rv#2 @gully
### Brittle TelFit Installation
Installation of the `telfit` code hinges on having the right fortran and gcc versions installed, and in some cases requires direct modification of the source code to get TelFit to compile, in turn making it difficult to install `igrins_rv`.  What upstream changes to TelFit could streamline the installation of `igrins_rv`?
