### 2021-04-05 shihyuntang/igrins_rv#3 @gully
TelFit Dependency Strategy
### Tradeoff of TelFit fork and future-proofing
As (bad?) luck would have it, the `fortranformat` dependency of TelFit was updated two days ago for the first time in seven years.
 This update introduced backwards-breaking changes and made the TelFit installation error during the `python setup.py install` phase.  I reported the issue in https://github.com/kgullikson88/Telluric-Fitter/issues/25 and made a pull request https://github.com/kgullikson88/Telluric-Fitter/pull/26.
The `igrins_rv` choice to include raw TelFit source code directly as part of this repository has some benefits and drawbacks.  On one hand you can control the exact behavior of TelFit and modify its code to work more closely with `igrins_rv`: deleterious changes to telfit are kept out.  On the other hand, you have to maintain your fork and do not effortlessly benefit from TelFit maintenance: positive changes to TelFit are kept out.  In this example, users of `igrins_rv` attempting to follow your installation instructions would encounter this `fortranformat` problem and not automatically benefit from the upstream fixes.
Changes to TelFit are small and infrequent at the moment---averaging 3 Pull Requests in ~4 years---so it is easy enough to make those changes manually.  The main drawback I see with the forking of  TelFit is to close off the development ecosystem.  A forked TelFit makes it ever-so-slightly harder for prospective developers of `igrins_rv` to interoperate: if I have a separate code that depends on the original TelFit, I cannot simultaneously use the two.

The current change to TelFit is a single line of code: changing the Default wavelength scaling from air to vacuum.  Is there a way to change this default in the execution call to TelFit rather than forking the master version?

### TelFit Installation woes act as a barrier to entry
Another TelFit-related thorn is that TelFit is hard to install.  The [TelFit Issues](https://github.com/kgullikson88/Telluric-Fitter/issues) have several instances of folks reporting installation woes.  Anecdotally, several researchers have personally asked or emailed me for help to install, or have reported abandoning altogether after TelFit installation failed.  I'm a contributor to TelFit and have a working version on my computer and still had a hard time reinstalling this version for `igrins_rv`.  I'm not sure what the solution is, since again that's more of an "upstream problem", but it does have consequences for `igrins_rv`--- TelFit currently acts as a barrier to entry for new users (especially students) and may limit the reach of the package.  In essence, what steps can be made to make it easier to install TelFit?
