### 2021-04-05 #3095 @gully
`Update`: I had trouble installing TelFit and highlighted some of the challenges and impact in https://github.com/shihyuntang/igrins_rv/issues/3. As part of this review it would be valuable to move away from packaging the TelFit source code with `igrins_rv` and instead pin to a version of the main TelFit repo.  I think this change can be accomplished without upstream changes.
