### 2021-04-05 shihyuntang/igrins_rv#2 @gully
## Running the code

### Quiet mode
Running the code prints out a lot of TelFit info to stdout.  It may be useful to enable a verbose mode or equivalently a quiet mode that replaces this text with a progress bar to help the user gauge how much time remains.  The text can be redirected to a console log if users want to inspect the instantaneous output.
