### 2021-04-05 shihyuntang/igrins_rv#3 @shihyuntang
Hi @gully ,

Thank you for bring this up!
The one line change in the Telfit source code is not a bug, but more like missing features(?). I just open an issue on Telfit kgullikson88/Telluric-Fitter#27 asking for "adding the option to change the status of vac2air from the call of the TelluricFitter class." After this feature been added in Telfit main repo, modification on `IGRINS RV` is straightforward.
