### 2021-04-05 shihyuntang/igrins_rv#2 @gully
### Overview and Workflow
The [Overview and Workflow documentation](https://github.com/shihyuntang/igrins_rv/wiki/Overview-and-Workflow) describes the order and conditions under which the various steps can be run. The conditional, iterative, and/or in-series nature of each step sequence takes some mental mapping when a newcomer is reading for the first time.  A diagram summarizing the key inputs and outputs of each step could make it easier to understand the workflow at a glance.
