### 2021-04-05 shihyuntang/igrins_rv#4 @gully
Transition from print to logging with a progress bar
Currently the code prints lots of progress information to standard out.  The user may wish to get more or less information depending on their experience level.  If possible, I recommend using [tqdm](https://tqdm.github.io/) to enable a progress bar for each of the main steps, so the user can gauge execution time at a glance.  The code already uses a logger---is it feasible to migrate the print statements to logging `.info()`?

Output from third party packages like TelFit could be suppressed and possibly/optionally sent to the logger.  For an example implementation see the lightkurve utility [suppress_stdout](https://github.com/lightkurve/lightkurve/blob/8d14f29892205b178a7c64e7faf03171f218358a/src/lightkurve/utils.py#L545).
