### 2021-04-05 #3095 @gully
`Update`: I got `igrins_rv` to run its steps 1 and 2 on the included example data.  I am now running step 3, which takes a while.  I have recommended the inclusion of a diagram of the workflow in the documentation.  I also recommend adding a progress bar, and transitioning from print to logging https://github.com/shihyuntang/igrins_rv/issues/4.
