### 2021-04-05 shihyuntang/igrins_rv#2 @gully
### FAQ
I recommend completing the [FAQ](https://github.com/shihyuntang/igrins_rv/wiki/FAQ) since I have a few of the same questions.
