### 2021-04-05 shihyuntang/igrins_rv#2 @gully
### Documentation
Most of the documentation resides in the wiki describing the core commandline interface (CLI) to run `igrins_rv` in three or four steps.  The documentation in the wiki pages is adequate to reproduce the example.  Most functions and classes in `Engine` lack docstrings.  These Engine functions are not intended to be directly called by the user.
