### 2021-04-05 shihyuntang/igrins_rv#2 @gully
### Testing
The main vehicle for testing appears to be the functional testing provided through running the scripts on example input data.  Running the code on this example data successfully produces RV values and uncertainties with scatter comparable to the 25 m/s claimed performance.   The individual functions within `Engine` do not have unit testing.
