### 2021-04-05 #3095 @gully
Overall the code appears to work as advertised for the example data.  The documentation and testing center on the commandline interface (CLI), and are adequate to reproduce the claimed results with the included example data.  I have not attempted to run the code on data beyond the examples.  The algorithmic ingredients (such as those in `Engine`) do not appear to be documented or tested.
