### 2021-04-08 #3095 @arfon
FYI, I just heard from the AAS publishing team that the DOI for their paper will be `10.3847/1538-3881/abf5e7`.
