### 2021-04-09 30617de
pass air2wave into GenerateModel
