### 2021-04-10 shihyuntang/igrins_rv#4 @shihyuntang
Hi @gully ,

Thank you for your suggestion. I tried the `suppress_stdout`, but was only able to remove some part of the printout.
Something like these will remain:
```
Note: The following floating-point exceptions are signalling: IEEE_UNDERFLOW_FLAG IEEE_DENORMAL
STOP  LBLRTM EXIT
0.48user 0.58system 0:01.07elapsed 99%CPU (0avgtext+0avgdata 7752maxresident)k
0inputs+37328outputs (0major+958minor)pagefaults 0swaps
Note: The following floating-point exceptions are signalling: IEEE_UNDERFLOW_FLAG IEEE_DENORMAL
STOP  LBLRTM EXIT
0.57user 0.48system 0:01.06elapsed 98%CPU (0avgtext+0avgdata 7780maxresident)k
0inputs+37064outputs (0major+959minor)pagefaults 0swaps

Running exec: lblrtm
```
I think this is because of different ways that Telfit is calling them?
Any further suggestion?
Thank you!
