+++ Telluric-Fitter-master/src/TelluricFitter.py
@@ -609,7 +609,7 @@ class TelluricFitter:
 
 
 
-    def GenerateModel(self, pars, nofit=False, separate_source=False, return_resolution=False, broaden=False, model=None, air_wave=True):
+    def GenerateModel(self, pars, nofit=False, separate_source=False, return_resolution=False, broaden=False, model=None):
         """
         This function does the actual work of generating a model with the given parameters,
         fitting the continuum, making sure the model and data are well aligned in
@@ -626,7 +626,6 @@ class TelluricFitter:
                         Ignored if nofit=False
         :param model: A DataStructures.xypoint instance containing an un-broadened telluric model.
                       If given, it uses this instead of making one.
-        :param air_wave:  Are the wavelengths in air wavelengths? Default is True.
 
         :return:  The best-fit telluric model, as a DataStructures.xypoint instance where the x-axis is
                  sampled the same as the data (so you should be able to directly divide the two). If
@@ -665,18 +664,12 @@ class TelluricFitter:
         lat = self.observatory["latitude"]
         alt = self.observatory["altitude"]
 
-        print(f'''
-        *********
-        *********
-        air_wave = {air_wave}
-        *********
-        *********
-        ''')
+
         #Generate the model:
         if model is None:
             model = self.Modeler.MakeModel(pressure, temperature, wavenum_start, wavenum_end, angle, h2o, co2, o3, n2o, co,
                                            ch4, o2, no, so2, no2, nh3, hno3, lat=lat, alt=alt, wavegrid=None,
-                                           resolution=None, vac2air=air_wave)
+                                           resolution=None, vac2air=False)
 
             #Save each model if debugging
             if self.debug and self.debug_level >= 5:
