+++ Engine/rebin_jv.py
@@ -60,7 +60,7 @@ def rebin_jv(Wold,Sold,Wnew,verbose):
     #Verify that new wavelength scale is a subset of old wavelength scale.
     if verbose == True:
         if (Wnew[0] < Wold[0]) or (Wnew[-1] > Wold[-1]):
-            print('New wavelength scale not subset of old.')
+            # print('New wavelength scale not subset of old.')
 
     #Select integration or interpolation depending on change in dispersion.
     if PSnew <= PSold:
