+++ main_step1.py
@@ -908,7 +908,7 @@ def mp_run(args, inparam, Nthreads, jerp, orders, nights, masterbeam):
     else:
         func = partial(MPinstB, args, inparam, jerp, orders)
 
-    outs = pqdm(np.arange(len(nights)), func, n_jobs=Nthreads)
+    outs = pqdm(np.arange(len(nights)), func, n_jobs=Nthreads, unit='night')
 
     # pool = mp.Pool(processes = Nthreads)
     # outs = pool.map(func, np.arange(len(nights)))
@@ -1130,7 +1130,7 @@ For H band RVs: We do not expect any systematic changes in the H band as the res
 
     print('B nods done! Halfway there! \n Now processing the A nods...')
     for jerp in range(len(orders)):
-        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order)    
+        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order)
         outs = mp_run(args, inparam, args.Nthreads, jerp, orders, nightsFinal,'A')
 
     print('\n')
