+++ main_step1.py
@@ -1125,12 +1125,12 @@ For H band RVs: We do not expect any systematic changes in the H band as the res
     # Run order by order, multiprocessing over nights within an order
     print('Processing the B nods first...')
     for jerp in range(len(orders)):
-        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order)
+        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order))
         outs = mp_run(args, inparam, args.Nthreads, jerp, orders, nightsFinal,'B')
 
     print('B nods done! Halfway there! \n Now processing the A nods...')
     for jerp in range(len(orders)):
-        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order)
+        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), order))
         outs = mp_run(args, inparam, args.Nthreads, jerp, orders, nightsFinal,'A')
 
     print('\n')
