+++ Telluric-Fitter-master/src/MakeModel.py
@@ -103,7 +103,7 @@ class ModelerException(Exception):
 class Modeler:
     def __init__(self, debug=False,
                  TelluricModelingDirRoot=DEFAULT_TELLURICMODELING,
-                 nmolecules=12):
+                 nmolecules=12, printoutshow=print_show):
         """
         Initialize a modeler instance
 
@@ -112,12 +112,14 @@ class Modeler:
                                         installation puts this in ~/.TelFit/
         :param nmolecules: The number of molecules to include in the telluric model. Probably don't change
                            this, and definitely don't increase it!
+        :param printoutshow:  Printing all fortran ourputs? Default is True.
         :return: None
         """
 
         Atmosphere = defaultdict(list)
         indices = {}
         self.debug = debug
+        self.printoutshow = printoutshow
         if not TelluricModelingDirRoot.endswith("/"):
             TelluricModelingDirRoot = TelluricModelingDirRoot + "/"
         if not 'rundir1' in os.listdir(TelluricModelingDirRoot):
@@ -328,7 +330,7 @@ class Modeler:
     def MakeModel(self, pressure=795.0, temperature=283.0, lowfreq=4000, highfreq=4600, angle=45.0, humidity=50.0,
                   co2=368.5, o3=3.9e-2, n2o=0.32, co=0.14, ch4=1.8, o2=2.1e5, no=1.1e-19, so2=1e-4, no2=1e-4, nh3=1e-4,
                   hno3=5.6e-4, lat=30.6, alt=2.1, wavegrid=None, resolution=None, save=False, libfile=None,
-                  vac2air=True, printoutshow=True):
+                  vac2air=True):
         """
         Here is the important function! All of the variables have default values,
           which you will want to override for any realistic use.
@@ -364,7 +366,6 @@ class Modeler:
                                saved file will be written to this filename. Should be a string
                                variable. Ignored if save==False
         :param vac2air:        If True (default), it converts the wavelengths from vacuum to air
-        :param printoutshow:  Printing all fortran ourputs? Default is True.
 
         :return:               DataStructures.xypoint instance with the telluric model. The x-axis
                                is in nanometers and the y-axis is in fractional transmission.
@@ -445,9 +446,9 @@ class Modeler:
                 #Run lblrtm
                 cmd = "cd " + TelluricModelingDir + ";sh runlblrtm_v3.sh"
                 try:
-                    if printoutshow:
+                    if self.printoutshow:
                         command = subprocess.check_call(cmd, shell=True)
-                    if not printoutshow:
+                    if not self.printoutshow:
                         command = subprocess.check_call(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
                 except subprocess.CalledProcessError:
                     raise subprocess.CalledProcessError("Error: Command '{}' failed in directory {}".format(cmd, TelluricModelingDir))
@@ -464,9 +465,9 @@ class Modeler:
         #Run lblrtm for the last time
         cmd = "cd " + TelluricModelingDir + ";sh runlblrtm_v3.sh"
         try:
-            if printoutshow:
+            if self.printoutshow:
                 command = subprocess.check_call(cmd, shell=True)
-            if not printoutshow:
+            if not self.printoutshow:
                 command = subprocess.check_call(cmd, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
         except subprocess.CalledProcessError:
             raise subprocess.CalledProcessError("Error: Command '{}' failed in directory {}".format(cmd, TelluricModelingDir))
