### 2021-04-10 shihyuntang/igrins_rv#4 @shihyuntang
Hi @gully ,

Thank you again for the suggestion.
I am able to suppress all the Telfit output by modifying the subprocess code in Telfit (see kgullikson88/Telluric-Fitter#27).
Now printouts from the main_step1.py looks like:

I will let you know when this had been applied to all steps.
