+++ environment.yml
@@ -17,3 +17,4 @@ dependencies:
   - pip:
     - pysynphot
     - more_itertools
+    - pqdm
