+++ main_step3.py
@@ -721,7 +721,7 @@ For H band RVs: We do not expect any systematic changes in the H band as the res
 
     # Run order by order, multiprocessing over nights within an order
     for jerp in range(len(orders)):
-        if not args.debug: print('Working on order {:02d}/{:02d} ({})'.format(int(jerp+1), len(orders), orders[jerp]))
+        if not args.debug: print('Working on order {} ({:02d}/{:02d})'.format(orders[jerp], int(jerp+1), len(orders)))
 
         func = partial(rv_MPinst, args, inparam, orders, jerp, trk, step2or3 )
         outs = pqdm(np.arange(len(nightsFinal)), func, n_jobs=args.Nthreads)
@@ -1074,7 +1074,6 @@ For H band RVs: We do not expect any systematic changes in the H band as the res
         logger.info('vsini results:       mean={:1.4f} km/s, std={:1.4f} km/s'.format(np.nanmean(vsinifinalCombined),
                                                                                 np.nanstd(vsinifinalCombined)))
 
-    print('\n')
     warning_r = log_warning_id(f'{outpath}/{args.targname}_{args.band}.log', start_time)
     if warning_r:
         print(f'''
