+++ main_step1.py
@@ -1138,7 +1138,7 @@ WARNING!! you got warning message during this run. Please check the log file und
           {outpath}/{args.targname}_{args.band}_A0Fits.log
 **********************************************************************************
 ''')
-
+    print('\n')
     if not args.debug: logger.addHandler(stream_hander)
     logger.info('A0 Fitting Done!')
 
