+++ main_step1.py
@@ -13,7 +13,7 @@ from Engine.detect_peaks import detect_peaks
 from Engine.crmask    import CRmasker
 #-------------------------------------------------------------------------------
 #-------------------------------------------------------------------------------
-
+logger = logging.getLogger(__name__)
 
 def MPinstB(args, inparam, jerp, orders, i):
     # Main function for A0 fitting that will be threaded over by multiprocessing
@@ -1020,7 +1020,6 @@ if __name__ == '__main__':
     outpath = f'./Output/{args.targname}_{args.band}/A0Fits'
 #-------------------------------------------------------------------------------
     # Handle logger
-    logger = logging.getLogger(__name__)
     if args.debug:
         logger.setLevel(logging.DEBUG)
     else:
