+++ Engine/importmodule.py
@@ -1,5 +1,6 @@
 # import all need modules...
 import sys, argparse, os, ast, re, logging, nlopt
+import logging.handlers
 from os      import listdir
 from os.path import isfile, join, isdir
 # Check python version
@@ -170,3 +171,57 @@ def read_prepdata(args):
         obs = np.array([obs0[n[:8]] for n in nightsFinal])
 
         return xbounddict, maskdict, tagsA, tagsB, jds, bvcs, nightsFinal, orders, obs
+
+# -------------------------------------------------------------
+#---- logger init ---
+
+def listener_configurer():
+    root = logging.getLogger()
+    h = logging.handlers.RotatingFileHandler('mptest.log', 'a', 300, 10)
+    f = logging.Formatter('%(asctime)s: %(module)s.py: %(levelname)s--> %(message)s')
+    h.setFormatter(f)
+    root.addHandler(h)
+
+# This is the listener process top-level loop: wait for logging events
+# (LogRecords)on the queue and handle them, quit when you get a None for a
+# LogRecord.
+def listener_process(queue, configurer):
+    configurer()
+    while True:
+        try:
+            record = queue.get()
+            if record is None:  # We send this as a sentinel to tell the listener to quit.
+                break
+            logger = logging.getLogger(record.name)
+            logger.handle(record)  # No level or filter logic applied - just do it!
+        except Exception:
+            import sys, traceback
+            print('Whoops! Problem:', file=sys.stderr)
+            traceback.print_exc(file=sys.stderr)
+
+
+# The worker configuration is done at the start of the worker process run.
+# Note that on Windows you can't rely on fork semantics, so each process
+# will run the logging configuration code when it starts.
+def worker_configurer(queue):
+    h = logging.handlers.QueueHandler(queue)  # Just the one handler needed
+    root = logging.getLogger()
+    formatter = logging.Formatter('%(asctime)s: %(module)s.py: %(levelname)s--> %(message)s')
+    root.addHandler(h)
+    # send all messages, for demo; no other level or filter logic applied.
+    root.setLevel(logging.DEBUG)
+
+# This is the worker process top-level loop, which just logs ten events with
+# random intervening delays before terminating.
+# The print messages are just so you know it's doing something!
+# def worker_process(queue, configurer):
+#     configurer(queue)
+#     name = multiprocessing.current_process().name
+#     print('Worker started: %s' % name)
+#     for i in range(10):
+#         time.sleep(random())
+#         logger = logging.getLogger(choice(LOGGERS))
+#         level = choice(LEVELS)
+#         message = choice(MESSAGES)
+#         logger.log(level, message)
+#     print('Worker finished: %s' % name)
