+++ Engine/IO_AB.py
@@ -14,6 +14,19 @@ import sys
 
 def partial_loader(inpath0,order):
 
+    '''
+    Access data arrays from reduced IGRINS fits files.
+    
+    Inputs: 
+    inpath0 : Path to file, including filename up to ".spec.fits" suffix
+    order   : Echelle order, as characterized by file index (as opposed to m number; for conversion between the two, see Stahl et al. 2021)
+    
+    Outputs:
+    wavelist : Wavelength solution from plp reduction (known to be imprecise)
+    fluxlist : Corresponding flux
+    s2nlist  : Corresponding signal to noise ratio
+    '''
+    
     hdulist = fits.open(inpath0+'.spec.fits')
     tbdata_w = hdulist[1].data
     tbdata_f = hdulist[0].data
@@ -30,6 +43,27 @@ def partial_loader(inpath0,order):
 
 def init_fitsread(path,kind,beam,night,order,tag,band,Ncuts=None):
 
+    '''
+    Fetches reduced spectrum from file, performs some basic quality checks and cleaning, and returns the results in a convenient format.
+    
+    Inputs: 
+    path  : Path to fits file up to some subdirectory (folder organization varies depending on whether loading target or telluric standard, separate A and Bs or combined
+    kind  : Specifies whether loading a telluric standard or target
+    beam  : Frame type (A, B, or combinedAB)
+    night : Date of observation in YYYYMMDD
+    order : Echelle order, as characterized by file index (as opposed to m number; for conversion between the two, see Stahl et al. 2021)
+    tag   : Number of observation, represented by last four digits in fits filename
+    band  : H or K band
+    Ncuts : List of format [L,R], where L specifies number of pixels to trim off left (low wavelength) end of spectrum, and R vice versa
+    
+    Outputs:
+    x     : Absolute pixel value (if spectra has edges trimmed, will begin at a nonzero number)
+    wave  : Wavelength solution from plp reduction (known to be imprecise)
+    s     : Corresponding flux
+    u     :  Corresponding uncertainty
+    
+    '''
+    
     if kind not in ['A0','target']:
         print('kind MUST BE "target" OR "A0", FORCE QUITTING!')
         print(breaker)
@@ -132,7 +166,9 @@ def init_fitsread(path,kind,beam,night,order,tag,band,Ncuts=None):
     return x,wave,s,u
 
 def airtovac(wave):
-    """Converts air wavelengths to vaccuum wavelengths returns a float or array of wavelengths
+    """
+    From PyAstronomy repository - https://pyastronomy.readthedocs.io/en/latest/index.html (Czela et al. 2019)
+    Converts air wavelengths to vaccuum wavelengths returns a float or array of wavelengths
     INPUTS:
       wave - Wavelengths in air, in Angstroms, float or array
     OUTPUTS:
@@ -150,6 +186,25 @@ def airtovac(wave):
     return newwave
 
 def setup_templates(logger, kind='synthetic', band='K', temperature=5000, logg=4.5):
+    
+    '''
+    Fetches static stellar and/or telluric templates from file.
+    
+    Inputs: 
+    logger      : Mechanism to keep updating log.txt 
+    kind        : Specifies kind of stellar template ('synthetic' = IGRINS RV team generated models, 'phoenix' = for IGRINS RV team only, 'user' = user provided template)
+    band        : H or K band
+    temperature : Effective temperature corresponding to stellar template
+    logg        : log(g) corresponding to stellar template
+    
+    Outputs:
+    watm    : Wavelength scale of static telluric template
+    satm    : Corresponding flux of static telluric template
[…]
