+++ Engine/clips.py
@@ -1,6 +1,18 @@
 import numpy as np
 
 def sigmaclip(l_return,l_ref,n):
+    '''
+    Sigma clips an array by values of another array.
+    
+    Inputs:
+    l_return : Array to be sigma-clipped
+    l_ref    : Array whose values determine the sigma clipping
+    n        : Number of standard deviations to clip above/below
+    
+    Outputs:
+    l : Sigma-clipped l_return
+    '''
+    
     go = True; l = l_return; lm = l_ref;
     while go == True:
         lenbefore = len(lm)
@@ -12,6 +24,19 @@ def sigmaclip(l_return,l_ref,n):
             return l
 
 def sigmaclip_above(l_return,l_ref,n):
+    
+    '''
+    Sigma clips an array by values of another array, but only high values, not low. Breaks input spectrum up into pieces and sigma clips each on its own.
+    
+    Inputs:
+    l_return : Array to be sigma-clipped
+    l_ref    : Array whose values determine the sigma clipping
+    n        : Number of standard deviations to clip above/below
+    
+    Outputs:
+    l_out : Sigma-clipped l_return
+    '''
+    
     go = True; l_out = np.array([]); l_base = np.array([]);
     while go == True:
         lenbefore = len(l_ref);
@@ -27,6 +52,19 @@ def sigmaclip_above(l_return,l_ref,n):
         l_out = np.array([]); l_base = np.array([]);
 
 def basicclip_above(l_return,l_ref,nslices):
+    
+    '''
+    Trims an array by upper outliers of another array. Breaks input spectrum up into pieces and clips each on its own.
+    
+    Inputs:
+    l_return : Array to be clipped
+    l_ref    : Array whose values determine the clipping
+    n        : Number of pieces to break spectrum into before clipping
+    
+    Outputs:
+    l_out : Sigma-clipped l_return
+    '''
+    
     n = 3
     s2nd  = np.sort(l_ref)
     s2ndd = s2nd[::-1]
@@ -48,7 +86,17 @@ def basicclip_above(l_return,l_ref,nslices):
 
 
 def singler(xin):
-    # Index selector for use during isolation of CRs and misfits after first run at optimization
+    
+    '''
+    Index selector for use during isolation of CRs and misfits after first run at optimization
+    
+    Inputs:
+    xin : Array of indices
+    
+    Outputs:
+    xout : xin without any consecutive values 
+    '''
+    # 
     xout = np.array([])
     for x in xin:
         if (x+1 in xin and x-1 in xin) or (x+1 in xin and x+2 in xin) or  (x-1 in xin and x-2 in xin):
@@ -58,7 +106,16 @@ def singler(xin):
     return np.array(xout,dtype=int)
 
 def multipler(xin):
-    # Index selector for use during isolation of CRs and misfits after first run at optimization
+    
+    '''
+    Index selector for use during isolation of CRs and misfits after first run at optimization
+    
+    Inputs:
+    xin : Array of indices
+    
+    Outputs:
+    xout : xin with only consecutive values 
+    '''
     xout = np.array([])
     for x in xin:
         if (x+1 in xin and x-1 in xin) or (x+1 in xin and x+2 in xin) or  (x-1 in xin and x-2 in xin):
