### 2021-04-12 5536151
Update contfit.py
