### 2021-04-12 7ae2962
Update detect_peaks.py
