+++ Engine/rebin_jv.py
@@ -64,7 +64,7 @@ def rebin_jv(Wold, Sold, Wnew, verbose, logger=None):
     PSnew = (Wnew[-1] - Wnew[0]) / (Nnew-1) #new pixel scale
 
     #Verify that new wavelength scale is a subset of old wavelength scale.
-    if (verbose == True) and (logger not None):
+    if (verbose == True) and (logger is not None):
         if (Wnew[0] < Wold[0]) or (Wnew[-1] > Wold[-1]):
             logger.warning('New wavelength scale not subset of old.')
 
