+++ Telluric-Fitter-master/src/MakeTape5.py
@@ -1,217 +1,0 @@
-"""
-    This file will read in the file 'ParameterFile', which is a human-readable
-      version of the input for LBLRTM (called TAPE5). It will convert that into
-      TAPE5 for use by LBLRTM. You should never need to call this directly!
-
-
-
-
-    This file is part of the TelFit program.
-
-    TelFit is free software: you can redistribute it and/or modify
-    it under the terms of the MIT license.
-
-    TelFit is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-
-    You should have received a copy of the MIT license
-    along with TelFit.  If not, see <http://opensource.org/licenses/MIT>.
-
-
-
-"""
-from __future__ import print_function, division
-
-import sys
-
-from fortranformat import FortranRecordWriter as writer
-from numpy import pi, cos, tan, arctan
-
-
-"""
-  This is a convenience function to convert directly between ParameterFile
-    and TAPE5. It is not used in TelFit, because we edit some of the pars
-    (see MakeModel.py).
-"""
-
-
-def Convert(parfile="ParameterFile", atmosphere=None, output="TAPE5"):
-    pars = ReadParFile(parfile)
-    WriteTape5(pars, atmosphere, output)
-
-
-"""
-  This function simply reads in the ParameterFile and saves
-    the parameters in a list. I should probably change this
-    to a dictionary at some point, but that is a lot of work...
-"""
-
-
-def ReadParFile(parameterfile="ParameterFile"):
-    infile = open(parameterfile)
-    lines = infile.readlines()
-    infile.close()
-
-    pars = []
-    for line in lines:
-        if not line.startswith("!") and not line.startswith("#"):
-            pars.append(line[24:].strip())
-
-    return pars
-
-
-"""
-  This function takes the output of ReadParFile (above),
-    and writes it to TAPE5. 
-"""
-
-
-def WriteTape5(pars, atmosphere=None, output="TAPE5"):
-    outfile = open(output, "w")
-
-    description = pars[0]
-    outfile.write(writer(("A%i" % min(80, len(description) + 1))).write(("$%s" % description, )) + "\n")
-
-    # Record 1.2
-    ihirac = int(pars[1])
-    ilblf4 = int(pars[2])
-    icntnm = int(pars[3])
-    iaersl = int(pars[4])
-    iemit = int(pars[5])
-    iscan = int(pars[6])
-    ifilter = int(pars[7])
-    iplot = int(pars[8])
-    itest = int(pars[9])
-    iatm = int(pars[10])
-    imrg = int(pars[11])
-    ilas = int(pars[12])
-    iod = int(pars[13])
-    ixsec = int(pars[14])
-    mpts = int(pars[15])
-    npts = int(pars[16])
-    outfile.write(writer("(10(4x,i1),3x,i2, 3(4x,i1),2(1x,i4))").write((
-        ihirac, ilblf4, icntnm, iaersl, iemit, iscan, ifilter, iplot, itest, iatm, imrg, ilas, iod, ixsec, mpts,
-        npts)) + "\n")
-
-    if icntnm == 6:
-        sys.exit("Error! icntnm = 6 not supported!")
-    if iemit == 2:
-        sys.exit("Error! iemit = 2 not supported!")
-
-    # Record 1.3
-    v1 = float(pars[17])
-    v2 = float(pars[18])
-    sample = float(pars[19])
-    dvset = float(pars[20])
-    alfal0 = float(pars[21])
[…]
