+++ Telluric-Fitter-master/src/telfit/__init__.py
@@ -1,4 +1,0 @@
-from TelluricFitter import TelluricFitter
-from MakeModel import Modeler
-import DataStructures
-import FittingUtilities
