+++ Telluric-Fitter-master/examples/library_creation/move_library_here.sh
@@ -1,6 +1,0 @@
-#After running the python script to make the library, you can move it here by executing this
-
-for fname in `cat Library_Files.dat`
-do
-  mv $fname .
-done
