+++ Telluric-Fitter-master/docs/MakeModel.rst
@@ -1,6 +1,0 @@
-API for MakeModel
-=================
-
-
-.. autoclass:: MakeModel.Modeler
-   :members: __init__, EditProfile, MakeModel
