+++ Telluric-Fitter-master/docs/Tutorial.rst
@@ -1,10 +1,0 @@
-TelFit Tutorial
-================
-
-After installation, there should a few usage examples in ~/.TelFit/examples. We go into a bit more detail on what the examples are doing here.
-
-.. toctree::
-   :maxdepth: 1
-
-   Modeling
-   Simple Fitting <fitting>

[…]
