+++ Engine/Telfitter.py
@@ -61,10 +61,10 @@ def wavefit(par0, dpar0):
     # opt.set_maxtime(1200) #seconds
     # Quit optimization based on relative change in output fit parameters between iterations.
     # Choosing smaller change tolerance than 1e-6 has demonstrated no improvement in precision.
-    opt.set_ftol_rel(1e-10)
+    opt.set_ftol_rel(1e-14)
     parfit = opt.optimize(par0)
 
-    parfit_floor = round_decimals_down(parfit, decimals=10)
+    parfit_floor = round_decimals_down(parfit, decimals=14)
     return parfit_floor
 
 #------------
