+++ Engine/opt.py
@@ -414,8 +414,8 @@ def optimizer(par0, dpar0, hardbounds_v_ip, fitobj, optimize):
     # opt.set_maxtime(1200) #seconds
     # Quit optimization based on relative change in output fit parameters between iterations.
     # Choosing smaller change tolerance than 1e-6 has demonstrated no improvement in precision.
-    opt.set_ftol_rel(1e-14)
+    opt.set_ftol_rel(1e-10)
     parfit = opt.optimize(par0)
 
-    parfit_floor = round_decimals_down(parfit, decimals=14)
-    return parfit_floor
+    # parfit_floor = round_decimals_down(parfit, decimals=14)
+    return parfit_parfitfloor
