### 2021-04-12 9cbe8e1
Revert "Update opt.py"
This reverts commit c34d89c58a87150893f31036fa9ea9ccdd1677e1.
