### 2021-04-12 e7edf65
Revert "floor to 7 digits"
This reverts commit 444b97af472f9c7dee2fccf79ed0fcc091bc76ae.
