+++ Engine/opt.py
@@ -9,21 +9,6 @@ from Engine.rebin_jv import rebin_jv
 import sys
 
 #-------------------------------------------------------------------------------
-def round_decimals_down(number:float, decimals:int=2):
-    """
-    Returns a value rounded down to a specific number of decimal places.
-    Borrowed from https://kodify.net/python/math/round-decimals/
-    """
-    if not isinstance(decimals, int):
-        raise TypeError("decimal places must be an integer")
-    elif decimals < 0:
-        raise ValueError("decimal places has to be 0 or more")
-    elif decimals == 0:
-        return np.floor(number)
-
-    factor = 10 ** decimals
-    return np.floor(number * factor) / factor
-
 def fmodel_chi(par,grad):
     '''
     Function to be optimized. Computes model spectrum and compares it with data to calculate reduced chisq.
@@ -417,11 +402,10 @@ def optimizer(par0, dpar0, hardbounds_v_ip, fitobj, optimize):
     opt.set_lower_bounds(lows)
     opt.set_upper_bounds(highs)
 
-    # opt.set_maxtime(1200) #seconds
+    maxruntime = 1200 #seconds
+    opt.set_maxtime(1200) #seconds
     # Quit optimization based on relative change in output fit parameters between iterations.
     # Choosing smaller change tolerance than 1e-6 has demonstrated no improvement in precision.
-    opt.set_ftol_rel(1e-8)
+    opt.set_ftol_rel(1e-10)
     parfit = opt.optimize(par0)
-
-    parfit_floor = round_decimals_down(parfit, 7)
-    return parfit_floor
+    return parfit
