### 2021-04-13 shihyuntang/igrins_rv#4 @shihyuntang
Updates:
Steps 1--3 are now showing with the progress bar, using the [parallel tqdm](https://pqdm.readthedocs.io/en/latest/readme.html) pkg. Printouts in the old version will only show under DeBug mode now.
Thank @gully again for this suggestion. The code looks more professional with these progress bar!
(latest stable version of this update is under the JOSS_review_updates branch)
