### 2021-04-13 shihyuntang/igrins_rv#3 @shihyuntang
Updates:
With the pull request (kgullikson88/Telluric-Fitter#30) on `Telfit` been merged, `IGRINS RV` is now no longer need to pack with the `Telfit` code.

As for the **TelFit Installation woes act as a barrier to entry**. Without modifying the `Telfit` code, the best we can do now is to make the installation instruction from source as detailed as possible. I added in the FAQ for [How to install the right gfortran version for Telfit?](https://github.com/shihyuntang/igrins_rv/wiki/FAQ#q-how-to-install-the-right-gfortran-version-for-telfit).

I would love to help you and Kevin on making Telfit pip installable! In that case, all I need to do is add Telfit to .yml list.
