### 2021-04-13 shihyuntang/igrins_rv#2 @gully
### Community

The JOSS review guidelines seek some steps to supporting a community :

> There should be clear guidelines for third-parties wishing to:

> - Contribute to the software
> - Report issues or problems with the software
> - Seek support

The README file currently discourages public contributions:
> This package is NOT yet ready for public use

I recommend the inclusion of the JOSS community guidelines to the README, replacing the public use warning.  The code appears mature enough that outside users could reasonably benefit from and contribute back to the application.  Contributors should be encouraged at the present stage, since they essentially act as beta testers.  Some further steps could be taken to lower the barrier to entry for new contributors (see the next topic).

### Roadmap, Open Issues, Feature Ideas
A project roadmap would go a long way to facilitate community involvement.  It allows users to see where the project is going, its known issues, and whether features-of-interest are on the horizon.  As part of this review, I recommend a project roadmap, either as a wiki page or---as can be common practice---a standalone GitHub Issue.  GitHub now allows selected issues to be "pinned" so the prominence of a roadmap can be denoted in this way, if preferred.
I also recommend the chronicling of known Issues and feature requests in standalone GitHub Issues.  GitHub Issues have a community building effect: the presence of a few issues and workaround discussions invites users to submit their own bug reports, and gives the maintainers a sense for how the software is being used in practice.  To catalyze thinking on these new issues and roadmap, here is a brainstorm of some "evergreen" topics that could have a standalone Issue, and if applicable could be discussed in a Roadmap.

1. Runtime Performance: What is the current runtime performance (in minutes per spectrum or whatever preferred unit), and what plans are there to decrease the runtime?  Are the runtime bottlenecks known, if so what are they, and if not is there a plan to investigate them?
2.  Precision Performance: The code claims a precision of 25 m/s, what sets that limit and is there any plan to increase precision further?  The precision limits may be the primary concern for some PRV practitioners and having a forum for interested parties to contribute questions or ideas could be synergistic, even if you believe you have already hit the systematic noise floor of the data themselves.
3.  Modulatity/API-ification: Currently the code can only be run as a command-line interface (CLI).  What are the prospects of making the code more modular so that third-party packages can make programmatic calls outside of the shell?
