+++ paper/paper.md
@@ -44,7 +44,7 @@ In the past decade, RVs have primarily been measured from spectra in the optical
 
 ``IGRINS RV`` is a pipeline built for extracting precision RVs from spectra taken with the Immersion GRating INfrared Spectrometer (IGRINS) spectrograph [@yuk10; @park14; @mace16; @mace18]. This pipeline is built on the forward-modeling methodology that was successfully applied to CSHELL and PHOENIX spectra in the past [@croc12]. Compared to RVs obtained via cross-correlation with stellar templates used by past studies, ``IGRINS RV`` gives three times higher RV precision, about 25--30 m/s, around narrowline stars in both H and K bands, shown by years of monitoring on two RV standard stars, GJ\ 281 and HD\ 26257. This improvement is because of the use of a more robust approach to wavelength calibration and better telluric modeling through the generation of high resolution synthetic templates from fits to A0 standard observations. ``IGRINS RV`` has also demonstrated its effectiveness in identifying orbiting companions by successfully recovering the planet-induced RV signals of HD\ 189733 and Tau\ Boo\ A. ``IGRINS RV`` lets users choose to obtain absolute RVs or relative RVs, depending on whether their priority is coarse RV characterization or more precise RV monitoring. The code extends the science capabilities of an already powerful spectrograph, which lacked a publicly available RV pipeline until now. It facilitates the detection and/or characterization of exoplanets, binary stars, star clusters, and moving group members, and it enables such studies to be done in a more precise and uniform way.
 
-``IGRINS RV`` requires that the ``igrins plp v2.2.0`` [@leegul17] and ``Telfit`` [@gull14] packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.
+``IGRINS RV`` makes use of the ``astropy`` [@astropy:2013; @astropy:2018] on handing sky coordinates and barycentric velocity correction, ``scipy`` [@2020SciPy] and ``numpy`` [@2020NumPy] on mathmatical calculation, ``nlopt`` [@john08; @box1965new] on the optimization process, ``pandas`` [@reback2020pandas; @mckinney-proc-scipy-2010]on data management, ``matplotlib`` [@Hunter2007] on plotting, and visulization, and `pqdm` on parallel progress bar visulization. ``IGRINS RV`` also requires that the ``igrins plp v2.2.0`` [@leegul17] and ``Telfit`` [@gull14] packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.
 
 
 
