+++ paper/master.bib
@@ -138,3 +138,23 @@ booktitle = {Ground-based and Airborne Instrumentation for Astronomy V},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2010SPIE.7735E..1MY},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@MISC{john08,
+   author = {{Johnson}, S.G.},
+    title = "{The NLopt nonlinear-optimization package}",
+    journal = {GitHub},
+    url = {https://github.com/stevengj/nlopt},
+     year = 2008
+}
+
+@article{box1965new,
+  title     = {A new method of constrained optimization and a comparison with other methods},
+  author    = {Box, MJ},
+  journal   = {The Computer Journal},
+  volume    = {8},
+  number    = {1},
+  pages     = {42--52},
+  year      = {1965},
+  publisher = {The British Computer Society},
+  doi       = {10.1093/comjnl/8.1.42}
+}
