### 2021-04-14 shihyuntang/igrins_rv#2 @shihyuntang
Hi @gully ,
Here are our reply to your comments above:

**Repository size**
    --> With removing the Telfit pkg and some unused data under `syn_template`, the repository is now weighs at 358 MB. We would like to keep the example data and synthetic templates in one place.

**Inclusion of raw TelFit code**
     --> TelFit source code removed from igrins rv repo.

**Citation of dependencies**
    --> A new paragraph has been added under the Statement of need:
> IGRINS RV makes use of the astropy (Astropy Collaboration et al., 2018, 2013) on handing sky coordinates and barycentric velocity correction, scipy (Virtanen et al., 2020) and numpy (Harris et al., 2020) on mathmatical calculation, nlopt (Box, 1965; Johnson, 2008) on  the optimization process, pandas (McKinney, 2010; team, 2020)on data management, and matplotlib (Hunter, 2007) on plotting. We also borrowed a part of code from BMC (Duarte & Watanabe, 2021) for peak detection. IGRINS RV requires that the igrins plp v2.2.0  (Lee et al., 2017) and Telfit (Gullikson et al., 2014) packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.

**Brittle TelFit Installation**
    --> Make Telfit able to be installed via pip should solve this issue. After that, we will only need to add Telfit in the .yml file for the installation.

**Quiet mode**
    --> done

**Overview and Workflow**
    --> Thank you for this suggestion. A **Workflow Diagram for IGRINS RV** has been added under [Overview and Workflow](https://github.com/shihyuntang/igrins_rv/wiki/Overview-and-Workflow) wiki page.

**FAQ**
    --> done.

**Documentation**
    --> Docstrings for functions under Engine are all been added.

**Testing**
    --> Currently, the main method of testing the code is by running the test data provided. We will consider to add function units test function in the future if needed.

**Community**
    --> Ways to Contribute to the software, Report issues or problems with the software, and Seek support have been added in the README.md as
> IGRINS RV is now under intense development. If you have any question, idea or wish to report any bug, please let us know by either open an issue or contact us (asa.stahl@rice.edu or sytang@lowell.edu). More on how to contribute can be found in the FAQ page.
On FAQ you can now find **How do I report bugs?** and **How do I help make IGRINS RV better?**.

As for the
> This package is NOT yet ready for public use
Thank you for bringing this up. Yes, we should remove this by now because the code is pretty much stable.

**Roadmap, Open Issues, Feature Ideas*
    --> I really appreciate this suggestion. For IGRINS RV being my first big project on Github, I have never taught of these before. An issue page https://github.com/shihyuntang/igrins_rv/#7 has been opened for ideas of future improvement.

**Precision Performance**
    --> We have added this in the FAQ to explain the limitation on the RV precision:
> IGRINS RV uses telluric (atmospheric) absorption lines as a common-path wavelength calibrator, so its precision is limited by the internal RV stability of Earth's atmosphere. This has been estimated to be ~10-20 m/s, so it is unlikely IGRINS RV will achieve better than the current ~25m/s precision. If you are obtaining worse precisions than this for your science targets and want to try to push toward better precisions, improving the accuracy of your stellar templates or increasing the signal quality of your data could help (note the precision achievable by IGRINS RV also depends on the vsin(i) of the target in question).

Thank you again for all these great suggestion.
Now the master branch is updated with the latest version of the code.
