### 2021-04-26 #3095 @gully
The actual paper portion should include comparisons to existing RV frameworks.  How does `igrins_rv` differ from those, and why are those tools unsuitable to the task, prompting the creation of `igrins_rv`?  For example, one such package that comes to mind is [wobble](https://github.com/megbedell/wobble).  One of wobble's limitations (if I recall correctly) is that it requires epochs acquired with a range of BERV's, and assumes somewhat isolated telluric lines, which may not adequately apply to the in-hand IGRINS data and the dense tellurics in the near-IR.  What other packages like wobble are out there, and why are they deemed to be unsuitable for the task?
Other packages like [radvel](https://github.com/California-Planet-Search/radvel) and [exoplanet](https://docs.exoplanet.codes/en/stable/) consume post-processed radial velocity velocities for science, but are not designed to do any of the custom pre-processing shown here; making that distinction may further clarify for readers the niche that `igrins_rv` satisfies.

At the moment, a prospective contributor to this codebase may face a challenge to extend or modify the code owing to 1) inherent complexity of precision echelle modeling, 2) limited standalone documentation for sub-routines, and 3) complexity arising from the interconnections among sub-routines (i.e. software architecture). Item 1 can't be changed, it's inherent to the problem!  Regarding item 2, the `igrins_rv` team has made some progress on standalone documentation by adding docstrings to some sub-functions.

Finally item 3 appears to be the hardest to describe and ameliorate: there is a delicate balance between solving the problem at hand and designing the code to be extensible to other related problems.  In essence my question is: to what extent could `igrins_rv` be further modularized?  `igrins_rv` laudably takes advantage of existing specialized codes such as `telfit` and the IGRINS-specific `plp`.  One step that could further simplify the code is to modularize the IGRINS-specific components. For example, what portions of the code are IGRINS-specific, and what portions can apply to other near-IR echelle spectrographs?  Asking that question throughout the codebase should reveal opportunities to make the code more modular.  Modularity makes it easier to maintain and sustain the codebase, and ultimately attracts more contributors and users.
I have attempted to modularize some of the IGRINS file Input/Output and other routine operations in an [experimental API](https://muler.readthedocs.io/en/latest/api.html) based on the IGRINS PLP output, just like `igrins_rv`.  It is currently unstable and not suitable as a dependency.  It's a proof-of-concept.  In the longer term, one could imagine separating out the hum-drum parts of the code with a significant code reduction:

`clean_spectrum = IGRINSSpectrum(filename).remove_nans().trim_edges().remove_outliers()`

A small point regarding the inclusion of a custom `detect_peaks` algorithm.  Why was this custom code chosen over the seemingly equivalent [scipy find_peaks](https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.find_peaks.html)?
