+++ paper/master.bib
@@ -1,3 +1,23 @@
+@ARTICLE{bede19,
+       author = {{Bedell}, Megan and {Hogg}, David W. and {Foreman-Mackey}, Daniel and {Montet}, Benjamin T. and {Luger}, Rodrigo},
+        title = "{WOBBLE: A Data-driven Analysis Technique for Time-series Stellar Spectra}",
+      journal = {\aj},
+     keywords = {atmospheric effects, methods: data analysis, planets and satellites: detection, stars: individual: 51 Pegasi, Barnard's Star, HD 189733, techniques: radial velocities, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = oct,
+       volume = {158},
+       number = {4},
+          eid = {164},
+        pages = {164},
+          doi = {10.3847/1538-3881/ab40a7},
+archivePrefix = {arXiv},
+       eprint = {1901.00503},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AJ....158..164B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @ARTICLE{croc12,
    author = {{Crockett}, C.~J. and {Mahmud}, N.~I. and {Prato}, L. and {Johns-Krull}, C.~M. and
 	{Jaffe}, D.~T. and {Hartigan}, P.~M. and {Beichman}, C.~A.},
