### 2021-04-27 #3095 @shihyuntang
Dear @gully ,

Here are our reply:
* The actual paper portion should include comparisons to existing RV frameworks … and why are they deemed to be unsuitable for the task?
--> Thank you for bring `wobble` to our attention. We added this paragraph in the "Statement of need":
> Current open source codes for extracting RVs, such as ``wobble`` [@bede19], only work with spectrographs that are highly stable and well-characterized in their wavelength scales, such as those built with precision RVs in mind. This is not the  case for IGRINS. ``IGRINS RV`` is also tailored to take into account specific aspects of the IGRINS instrument, like the variations in spectral resolution across the detector and the year-long K band detector defocus.

* Other packages like radvel and exoplanet consume …  may further clarify for readers the niche that igrins_rv satisfies.
--> As our package's name: IGRINS RV, we think it is obvious that this is a package specifically for IGRINS instrument and is for RV reduction. Moreover, on the github README page as well as the front page of the wiki, we do direct reader/user to the AJ paper. In the AJ paper, we detailed the structure of the IGRINS code. Therefore, confusions is unlikely to happen.

* At the moment … separating out the hum-drum parts of the code with a significant code reduction:
--> We agree that IGRINS RV lacks modularity and that this should be improved in the future if needed. For now, we are only focusing on delivering robust RVs for the IGRINS spectrograph. The wide variations between different spectrographs (e.g. differences in characteristic instrumental broadening profile shape, resolution changes across the detector, blaze shape, wavelength stability) would make altering IGRINS RV to apply to another instrument not straightforward by any means, and we don't think it would be useful to provide tools for the user to attempt such a task. As the developers, we would be happy to extend IGRINS RV's capabilities to another instrument, but we would require a large amount of data (e.g. on RV standard stars) to do so. We will add a notice on the github page laying this out and provisionally offering to extend IGRINS RV to other spectrographs if users have such data and their spectrograph would be well-suited by our code, meaning it's NIR, echelle, and not designed for precision RV experiments (e.g., iSHELL).

--> The detect_peaks function under detect_peaks.py is adopted in the very early stage of the code development when we were not aware of the similar function in scipy. It works fine so we do not change it in the end.

Thank you.
