### 2021-05-10 #3095 @gully
I will double-check tomorrow any remaining tasks in this review.  I have not looked at the parallel paper linked in a previous comment thread.  Maybe a question to the editors---how should parallel papers be handled?
