### 2021-05-11 #3095 @gully
Looking over the PDF, the text appears to lack some references for how this package compares to other RV packages.  A comment above shows that a reference to `wobble` was added, but I cannot see that in the newly generated pdf.  Wobble is not the only relevant package though, and I am recommending that other comparable packages are cited to clarify how `igrins_rv` fits into the broader ecosystem of RV packages.  If there is truly such a dearth of open source packages, that can be explained with citations to papers employing closed-source packages.  I am also recommending the dependencies such as `numpy` and others be included unless the editors provide guidance to the contrary.
