+++ paper/paper.md
@@ -42,12 +42,15 @@ In the past decade, RVs have primarily been measured from spectra in the optical
 
 # Statement of need
 
-``IGRINS RV`` is a pipeline built for extracting precision RVs from spectra taken with the Immersion GRating INfrared Spectrometer (IGRINS) spectrograph [@yuk10; @park14; @mace16; @mace18]. This pipeline is built on the forward-modeling methodology that was successfully applied to CSHELL and PHOENIX spectra in the past [@croc12]. Compared to RVs obtained via cross-correlation with stellar templates used by past studies, ``IGRINS RV`` gives three times higher RV precision, about 25--30 m/s, around narrowline stars in both H and K bands, shown by years of monitoring on two RV standard stars, GJ\ 281 and HD\ 26257. This improvement is because of the use of a more robust approach to wavelength calibration and better telluric modeling through the generation of high resolution synthetic templates from fits to A0 standard observations. Current open source codes for extracting RVs, such as ``wobble`` [@bede19], only work with spectrographs that are highly stable and well-characterized in their wavelength scales, such as those built with precision RVs in mind. This is not the  case for IGRINS. ``IGRINS RV`` is also tailored to take into account specific aspects of the IGRINS instrument, like the variations in spectral resolution across the detector and the year-long K band detector defocus.
+Current RV pipelines and techniques that can deliver RV precision in tens of m/s (or beyond) in the NIR, e.g., ``PySHELL`` [@cale19], ``wobble`` [@bede19] or the PCA-based cross-correlating method used for the SPIRou spectrograph [@mout20], all require instruments that are highly stabilized and with well-characterized wavelength solutions. For example, the iSHELL spectrograph can be equipped with the methane isotopologue gas cell, and the SPIRou spectrograph comes with uranium-neon hollow-cathode lamps and stabilized Fabry-Perot etalons.
+
+The Immersion GRating INfrared Spectrometer (IGRINS) spectrograph [@yuk10; @park14; @mace16; @mace18], on the other hand, was not designed to be RV-stable and comes with no means of wavelength calibration accurate enough to achieve RVs precise to tens of m/s.  A new approach to extract precision RVs is needed for an echelle spectrograph like IGRINS, which still offers fertile ground for RV science with its high resolution (R ~ 45,000) and broad spectral grasp (the full H and K bands).
+
+``IGRINS RV`` is a pipeline tailored for extracting precision RVs from spectra taken with IGRINS on different facilities. This pipeline is built on the forward-modeling methodology that was successfully applied to CSHELL and PHOENIX spectra [@croc12] that utilized telluric lines as a common-path wavelength calibrator. Compared to RVs obtained by cross-correlation with stellar templates adopted by past studies, ``IGRINS RV`` gives three times higher RV precision, about 25--30 m/s, around narrow-line stars in both H and K bands, shown by years of monitoring on two RV standard stars, GJ\ 281 and HD\ 26257. ``IGRINS RV`` also pushes this technique, using telluric lines as wavelength calibrator for RV calculation, to its limits as studies found the stability of the telluric lines is about 10--20 m/s. Moreover, ``IGRINS RV`` is also tailored to take into account specific aspects of the IGRINS instrument, like the variations in spectral resolution across the detector and the year-long K band detector defocus.
 
[…]
