+++ paper/master.bib
@@ -351,3 +351,60 @@ archivePrefix = {arXiv},
   doi          = {10.5281/zenodo.4599319},
   url          = {https://doi.org/10.5281/zenodo.4599319}
 }
+
+
+@ARTICLE{gao16,
+       author = {{Gao}, Peter and {Plavchan}, P. and {Gagn{\'e}}, J. and {Furlan}, E. and {Bottom}, M. and {Anglada-Escud{\'e}}, G. and {White}, R. and {Davison}, C.~L. and {Beichman}, C. and {Brinkworth}, C. and {Johnson}, J. and {Ciardi}, D. and {Wallace}, K. and {Mennesson}, B. and {von Braun}, K. and {Vasisht}, G. and {Prato}, L. and {Kane}, S.~R. and {Tanner}, A. and {Crawford}, T.~J. and {Latham}, D. and {Rougeot}, R. and {Geneser}, C.~S. and {Catanzarite}, J.},
+        title = "{Retrieval of Precise Radial Velocities from Near-infrared High-resolution Spectra of Low-mass Stars}",
+      journal = {\pasp},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2016,
+        month = oct,
+       volume = {128},
+       number = {968},
+        pages = {104501},
+          doi = {10.1088/1538-3873/128/968/104501},
+archivePrefix = {arXiv},
+       eprint = {1603.05997},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016PASP..128j4501G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{cale19,
+       author = {{Cale}, Bryson and {Plavchan}, Peter and {LeBrun}, Danny and {Gagn{\'e}}, Jonathan and {Gao}, Peter and {Tanner}, Angelle and {Beichman}, Charles and {Xuesong Wang}, Sharon and {Gaidos}, Eric and {Teske}, Johanna and {Ciardi}, David and {Vasisht}, Gautam and {Kane}, Stephen R. and {von Braun}, Kaspar},
+        title = "{Precise Radial Velocities of Cool Low-mass Stars with iSHELL}",
+      journal = {\aj},
+     keywords = {atmospheric effects, infrared: stars, methods: data analysis, stars: individual: 61 Cygni A, Barnard{\textquoteright}s Star, GJ 15 A, techniques: radial velocities, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = nov,
+       volume = {158},
+       number = {5},
+          eid = {170},
+        pages = {170},
+          doi = {10.3847/1538-3881/ab3b0f},
+archivePrefix = {arXiv},
+       eprint = {1908.07560},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AJ....158..170C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{hira20,
[…]
