+++ paper/master.bib
@@ -408,3 +408,40 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020PASJ...72...93H},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@ARTICLE{mout20,
+       author = {{Moutou}, C. and {Dalal}, S. and {Donati}, J. -F. and {Martioli}, E. and {Folsom}, C.~P. and {Artigau}, {\'E}. and {Boisse}, I. and {Bouchy}, F. and {Carmona}, A. and {Cook}, N.~J. and {Delfosse}, X. and {Doyon}, R. and {Fouqu{\'e}}, P. and {Gaisn{\'e}}, G. and {H{\'e}brard}, G. and {Hobson}, M. and {Klein}, B. and {Lecavelier des Etangs}, A. and {Morin}, J.},
+        title = "{Early science with SPIRou: near-infrared radial velocity and spectropolarimetry of the planet-hosting star HD 189733}",
+      journal = {\aap},
+     keywords = {planetary systems, stars: magnetic field, stars: activity, techniques: radial velocities, techniques: polarimetric, techniques: spectroscopic, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2020,
+        month = oct,
+       volume = {642},
+          eid = {A72},
+        pages = {A72},
+          doi = {10.1051/0004-6361/202038108},
+archivePrefix = {arXiv},
+       eprint = {2008.05411},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...642A..72M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{zech18,
+       author = {{Zechmeister}, M. and {Reiners}, A. and {Amado}, P.~J. and {Azzaro}, M. and {Bauer}, F.~F. and {B{\'e}jar}, V.~J.~S. and {Caballero}, J.~A. and {Guenther}, E.~W. and {Hagen}, H. -J. and {Jeffers}, S.~V. and {Kaminski}, A. and {K{\"u}rster}, M. and {Launhardt}, R. and {Montes}, D. and {Morales}, J.~C. and {Quirrenbach}, A. and {Reffert}, S. and {Ribas}, I. and {Seifert}, W. and {Tal-Or}, L. and {Wolthoff}, V.},
+        title = "{Spectrum radial velocity analyser (SERVAL). High-precision radial velocities and two alternative spectral indicators}",
+      journal = {\aap},
+     keywords = {methods: data analysis, techniques: radial velocities, techniques: spectroscopic, planets and satellites: detection, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2018,
+        month = jan,
+       volume = {609},
+          eid = {A12},
+        pages = {A12},
+          doi = {10.1051/0004-6361/201731483},
+archivePrefix = {arXiv},
+       eprint = {1710.10114},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018A&A...609A..12Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
