+++ paper/paper.md
@@ -48,7 +48,7 @@ Current RV pipelines and techniques that can deliver RV precision in tens of m/s
 
 ``IGRINS RV`` has demonstrated its effectiveness in identifying orbiting companions by successfully recovering the planet-induced RV signals of HD\ 189733 and Tau\ Boo\ A. ``IGRINS RV`` lets users choose to obtain absolute RVs or relative RVs, depending on whether their priority is coarse RV characterization or more precise RV monitoring. The code extends the science capabilities of an already powerful spectrograph, which lacked a publicly available RV pipeline until now. It facilitates the detection and/or characterization of exoplanets, binary stars, star clusters, and moving group members, and it enables such studies to be done in a more precise and uniform way.
 
-``IGRINS RV`` makes use of the ``astropy`` [@astropy2013; @astropy2018] on handing sky coordinates and barycentric velocity correction, ``scipy`` [@2020SciPy] and ``numpy`` [@2020NumPy] on mathmatical calculation, ``nlopt`` [@john08; @box1965new] on the optimization process, ``pandas`` [@reback2020pandas; @mckinney2010]on data management, and ``matplotlib`` [@Hunter2007] on plotting. We also borrowed a part of code from `BMC` [@marcos2021] for peak detection. ``IGRINS RV`` requires that the ``igrins plp v2.2.0`` [@leegul17] and ``Telfit`` [@gull14] packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.
+``IGRINS RV`` makes use of the ``astropy`` [@astropy2013; @astropy2018] on handing sky coordinates and barycentric velocity correction, ``scipy`` [@2020SciPy] and ``numpy`` [@2020NumPy] on mathmatical calculation, ``nlopt`` [@john08; @box1965new] on the optimization process, ``pandas`` [@reback2020pandas; @mckinney2010] on data management, and ``matplotlib`` [@Hunter2007] on plotting. We also used a part of code from `BMC` [@marcos2021] for peak detection. ``IGRINS RV`` requires that the ``igrins plp v2.2.0`` [@leegul17] and ``Telfit`` [@gull14] packages be pre-installed. Detailed documentation and tutorials can be found on the GitHub wiki page.
 
 
 # Acknowledgements
