+++ paper/master.bib
@@ -445,3 +445,39 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2018A&A...609A..12Z},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{seif08,
+       author = {{Seifahrt}, A. and {K{\"a}ufl}, H.~U.},
+        title = "{High precision radial velocity measurements in the infrared. A first assessment of the RV stability of CRIRES}",
+      journal = {\aap},
+     keywords = {instrumentation: spectrographs, instrumentation: adaptive optics, methods: observational, techniques: radial velocities, techniques: spectroscopic, stars: AGB and post-AGB, Astrophysics},
+         year = 2008,
+        month = dec,
+       volume = {491},
+       number = {3},
+        pages = {929-939},
+          doi = {10.1051/0004-6361:200810174},
+archivePrefix = {arXiv},
+       eprint = {0809.1435},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2008A&A...491..929S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{fig10,
+       author = {{Figueira}, P. and {Pepe}, F. and {Lovis}, C. and {Mayor}, M.},
+        title = "{Evaluating the stability of atmospheric lines with HARPS}",
+      journal = {\aap},
+     keywords = {atmospheric effects, instrumentation: spectrographs, methods: observational, techniques: radial velocities, planetary systems, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2010,
+        month = jun,
+       volume = {515},
+          eid = {A106},
+        pages = {A106},
+          doi = {10.1051/0004-6361/201014005},
+archivePrefix = {arXiv},
+       eprint = {1003.0541},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010A&A...515A.106F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
