+++ paper/paper.md
@@ -37,12 +37,12 @@ bibliography: master.bib
 
 The relative radial velocity of a star with respect to the Sun can be calculated from its electromagnetic spectrum using the Doppler Effect. This line-of-sight motion, called the Radial Velocity (RV), is an essential tool for astrophysicists. RVs are not only used to detect and characterize exoplanets, but also play a key role in studies of binary stars, star clusters, and moving group member identification.
 
-In the past decade, RVs have primarily been measured from spectra in the optical wavelength regime. This is partly because of advancements in detector technology, but also because of the paucity of Earth's atmospheric absorption features (telluric lines) in the optical. Yet for a fainter, cooler, smaller stellar object like an M type star (stars with mass less than half of the Sun), which emits more energy in the Near-Infrared (NIR), observations in the NIR can save a considerable amount of exposure time. Also, the M type star is the most common type of star. This along with its size increases the detectability of Earth-like planets around them. Moreover, the stellar activity that can drive false positive exoplanet detections, e.g., star spots carried into view by stellar rotation is shown to be less severe in the NIR compared to optical.
+In the past decade, RVs have primarily been measured from spectra in the optical wavelength regime. This is partly because of advancements in detector technology, but also because of the paucity of Earth's atmospheric absorption features (telluric lines) in the optical. Yet for a fainter, cooler, smaller stellar object like an M type star (stars with mass less than half of the Sun), which emits more energy in the Near-Infrared (NIR), observations in the NIR can save a considerable amount of exposure time. Also, the M type star is the most common type of star. This along with its size increases the detectability of Earth-like planets around them. Moreover, the stellar activity that can drive false positive exoplanet detections, e.g., star spots carried into view by stellar rotation, is shown to be less severe in the NIR compared to optical.
 
 
 # Statement of need
 
-Current RV pipelines and techniques that can deliver RV precision in tens of m/s (or better) in the NIR, e.g., ``PySHELL`` [@cale19], ``wobble`` [@bede19], ``SERVAL`` [@zech18] or the PCA-based cross-correlating method used for the SPIRou spectrograph [@mout20], all require instruments that are highly stabilized and have well-characterized wavelength solutions. For example, the iSHELL spectrograph can be equipped with the methane isotopologue gas cell, and the SPIRou and the CARMENES (NIR channel) spectrograph come with uranium-neon hollow-cathode lamps and stabilized Fabry-Perot etalons. The Immersion GRating INfrared Spectrometer (IGRINS) spectrograph [@yuk10; @park14; @mace16; @mace18], on the other hand, was not designed to be RV-stable and comes with no means of wavelength calibration accurate enough to achieve RVs precise to tens of m/s using existing techniques. A new approach to extract precision RVs is needed for an echelle spectrograph like IGRINS, which offers fertile ground for RV science with its high resolution (R ~ 45,000) and broad spectral grasp (the full H and K bands).
[…]
