### 2021-05-17 shihyuntang/igrins_rv#3 @shihyuntang
Hi @gully ,

I made a [pull request](https://github.com/kgullikson88/Telluric-Fitter/pull/32) on `Telfit` that fix the error while compiling the  LBLRTM with `gfortran` (see [this](https://github.com/kgullikson88/Telluric-Fitter/issues/24#issuecomment-840932473) issue on `Telfit` for more details). If this fix can be verified by you or Kevin, making `Telfit` pip installable with py3 will help a lot for `IGRINS RV` --- overcome the **TelFit Installation woes act as a barrier to entry** problem
