+++ paper/master.bib
@@ -1,7 +1,7 @@
 @ARTICLE{bede19,
        author = {{Bedell}, Megan and {Hogg}, David W. and {Foreman-Mackey}, Daniel and {Montet}, Benjamin T. and {Luger}, Rodrigo},
         title = "{WOBBLE: A Data-driven Analysis Technique for Time-series Stellar Spectra}",
-      journal = {\aj},
+      journal = {Astronomical Journal},
      keywords = {atmospheric effects, methods: data analysis, planets and satellites: detection, stars: individual: 51 Pegasi, Barnard's Star, HD 189733, techniques: radial velocities, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2019,
         month = oct,
@@ -22,7 +22,7 @@ archivePrefix = {arXiv},
    author = {{Crockett}, C.~J. and {Mahmud}, N.~I. and {Prato}, L. and {Johns-Krull}, C.~M. and
 	{Jaffe}, D.~T. and {Hartigan}, P.~M. and {Beichman}, C.~A.},
     title = "{A Search for Giant Planet Companions to T Tauri Stars}",
-  journal = {\apj},
+  journal = {Astrophysical Journal},
 archivePrefix = "arXiv",
    eprint = {1211.1389},
  primaryClass = "astro-ph.EP",
@@ -40,7 +40,7 @@ archivePrefix = "arXiv",
 @ARTICLE{gull14,
        author = {{Gullikson}, Kevin and {Dodson-Robinson}, Sarah and {Kraus}, Adam},
         title = "{Correcting for Telluric Absorption: Methods, Case Studies, and Release of the TelFit Code}",
-      journal = {\aj},
+      journal = {Astronomical Journal},
      keywords = {atmospheric effects, instrumentation: spectrographs, techniques: spectroscopic, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2014,
         month = sep,
@@ -79,7 +79,7 @@ archivePrefix = {arXiv},
     title = "{300 nights of science with IGRINS at McDonald Observatory}",
 booktitle = {Ground-based and Airborne Instrumentation for Astronomy VI},
      year = 2016,
-   series = {\procspie},
+   series = {Proceedings of the SPIE},
    volume = 9908,
     month = aug,
       eid = {99080C},
@@ -103,7 +103,7 @@ booktitle = {Ground-based and Airborne Instrumentation for Astronomy VI},
          {Levine}, Stephen and {Nofi}, Larissa and {Lopez-Valdivia}, Ricardo and
          {Weinberger}, Alycia J. and {Jaffe}, Daniel T.},
         title = "{IGRINS at the Discovery Channel Telescope and Gemini South}",
-    booktitle = {\procspie},
+    booktitle = {Proceedings of the SPIE},
          year = 2018,
        series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
        volume = {10702},
@@ -128,7 +128,7 @@ booktitle = {Ground-based and Airborne Instrumentation for Astronomy VI},
     title = "{Design and early performance of IGRINS (Immersion Grating Infrared Spectrometer)}",
 booktitle = {Ground-based and Airborne Instrumentation for Astronomy V},
      year = 2014,
-   series = {\procspie},
+   series = {Proceedings of the SPIE},
    volume = 9147,
     month = jul,
       eid = {91471D},
[…]
