### 2021-06-07 #3095 @shihyuntang
Thank you @jmbrewer !

Two questions for @dfm :
1. Is the pdf I just asked whedon to generate serves as the *proof* as other traditional journals do?
2. About the `Telfit` package installation issue (a pkg used in `IGRINS RV`). I am still waiting for its author (kgullikson88) to merge my pull request [here](https://github.com/kgullikson88/Telluric-Fitter/pull/32) and make a new release of pip. If this is done, we will no longer need to manually source install (+ several changes in the source files) `Telfit` but just add `pip install Telfit` in the .yml file. However, I haven't heard from kgullikson88 for a while. Do you think I should wait so that the JOSS publish version has this fix? Or can I publish the current version of `IGRINS RV` on JOSS and make a new release on Github after `Telfit` is pip insatiable? @gully can you help on this since you reviewed my pull request on `Telfit`.

Thank you!
