### 2021-06-07 #3095 @dfm
> Is the pdf I just asked whedon to generate serves as the proof as other traditional journals do?

Yes! There will possibly be one final round of edits later in the process, but now is the time to check for any final issues.
