### 2021-06-08 #3095 @gully
OK @shihyuntang `telfit` has `pip install` working for python 3, including the change you made for the setup.py configuration.  Thank you!  You have one open PR over there ("resolution2") that that I think I can merge next.
