### 2021-06-08 #3095 @shihyuntang
Hi @gully
Wow, thank you!
My other pull request ("resolution2") for `Telfit` is not only changing the resolution --> resolution2, but also fixed the print statements for example Fit.py file to let it be py3.X compatible.
(I have fixed the conflict for that branch. Now you should be able to just pull it.)
