### 2021-06-09 #3095 @shihyuntang
Dear @dfm ,

> Comment @whedon generate pdf on this thread and read through the manuscript to make sure that you're happy with it (it's hard to make changes later!), especially author names and affiliations.

-->Done

> Increment the version number of the software and report that version number back here.

--> v1.0.0

> Create an archived release of that version of the software (using Zenodo or something similar). Please make sure that the metadata (title and author list) exactly match the paper. Then report the DOI of the release back to this thread.

--> DOI: 10.5281/zenodo.4917291. On [https://zenodo.org/record/4917291#.YMDuNC_fC-Y](https://zenodo.org/record/4917291#.YMDuNC_fC-Y)

Thank you!
